:recycle: (pc-11984) deprecate BicolorLocationPointer icon
